feat: refactor login and register page
