+++ client/src/pages/authentication/login/login.tsx
@@ -9,13 +9,13 @@ import {
   Typography,
 } from '@material-ui/core';
 import { FaFacebookF, FaGoogle } from 'react-icons/fa';
-import { useStyles } from '../../styles/authentication/login';
+import { useStyles } from './loginStyle';
 import { Formik } from 'formik';
 import React, { useState, useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
 import { RouteComponentProps, Link } from 'react-router-dom';
 import * as yup from 'yup';
-import { LoginInput, useLoginMutation } from '../../graphql';
+import { LoginInput, useLoginMutation } from '../../../graphql';
 
 type DataKeys = keyof LoginInput;
 type Data = { [key in DataKeys]: string };
