+++ client/src/pages/authentication/register/registerStyle.ts
