+++ e2e/cypress/integration/authentication/register.ts
@@ -1,4 +1,4 @@
-import register from '../support/authentication/RegisterPage';
+import register from '../../support/authentication/RegisterPage';
 
 describe('Register', () => {
   const RegisterPage = register('/register');
