--- server/database/docker-compose.db.yml
@@ -1,9 +1,0 @@
-version: '3'
-services:
-  db:
-    container_name: re-radio_db
-    image: postgres:10.7
-    restart: "no"
-    environment:
-      POSTGRES_USER: ${DB_USER}
-      POSTGRES_PASSWORD: ${DB_PASSWORD}
